Fix "approve" routes to account for missing rows
For approvals, missing rows, or "no data", does not mean anything other than the workflow did not require any approvals or the decision is yet to be completed by the user. Most other fields return data from the database, however, approval returns a "nodata" status when the data is missing in the database.