rename returning resources from annotation
Resources in the hub cluster might have to be renamed due to the nature of sharing one hub cluster for multiple remote clusters. Resources, namely secrets, with the same name would cause data mix-ups and or other potential issues.

Upon entering the cluster, outputsSecret names may be changed to work on the hub. When this is done, an annotation can be used on the tf resource that will preserve the original name. When the API returns the secret to the remote, it changes the name of the secret on the way out to the configured one in the annotation.