refactoring tfo-unlock branch
Changes any tfo-unlock references to terraform-unlock for clarity, (tfo doesn't have any locking mechanisms).
Remove the conditionals in the New function. This introduces some redundancy but simplifies code readability.

For the most part, happy paths should go ended at the end of functions and should not return the happy path in the middle of a function.